+++ core/server/api_container/main.go
@@ -90,9 +90,9 @@ func runMain() error {
 		return stacktrace.NewError("Kurtosis backend type is '%v' but cluster configuration parameters are null.", args.KurtosisBackendType_Kubernetes.String())
 	}
 
-	gitModuleManager, err := enclaveDataDir.GetGitModuleManager()
+	gitModuleContentProvider, err := enclaveDataDir.GetGitModuleContentProvider()
 	if err != nil {
-		return stacktrace.Propagate(err, "An error occurred while creating the git module manager")
+		return stacktrace.Propagate(err, "An error occurred while creating the Git module content provider")
 	}
 
 	// TODO Extract into own function
@@ -150,8 +150,8 @@ func runMain() error {
 	}()
 
 	// TODO: Consolidate Interpreter, Validator and Executor into a single interface
-	startosisInterpreter := startosis_engine.NewStartosisInterpreter(serviceNetwork, gitModuleManager)
 	startosisValidator := startosis_engine.NewStartosisValidator(&kurtosisBackend)
+	startosisInterpreter := startosis_engine.NewStartosisInterpreter(serviceNetwork, gitModuleContentProvider)
 	startosisExecutor := startosis_engine.NewStartosisExecutor()
 
 	//Creation of ApiContainerService
@@ -192,7 +192,7 @@ func createServiceNetworkAndModuleStore(
 	enclaveDataDir *enclave_data_directory.EnclaveDataDirectory,
 	args *args.APIContainerArgs,
 	ownIpAddress net.IP,
-) (*service_network.ServiceNetwork, *module_store.ModuleStore, error) {
+) (service_network.ServiceNetwork, *module_store.ModuleStore, error) {
 	enclaveIdStr := args.EnclaveId
 	enclaveId := enclave.EnclaveID(enclaveIdStr)
 
@@ -224,7 +224,7 @@ func createServiceNetworkAndModuleStore(
 		kurtosisBackend,
 		enclaveId)
 
-	serviceNetwork := service_network.NewServiceNetwork(
+	serviceNetwork := service_network.NewDefaultServiceNetwork(
 		enclaveId,
 		ownIpAddress,
 		args.GrpcListenPortNum,